update help widget

+++ src/draw.rs
@@ -364,7 +364,7 @@ fn draw_summary<B: Backend>(frame: &mut Frame<B>, app: &mut App, mut area: Rect)
 fn draw_help<B: Backend>(frame: &mut Frame<B>, app: &mut App, area: Rect) {
     let mut layout = area;
 
-    if layout.width < HELP_WIDTH || layout.height < HELP_HEIGHT {
+    if layout.width < HELP_WIDTH as u16 || layout.height < HELP_HEIGHT as u16 {
         frame.render_widget(
             Paragraph::new(Text::styled(
                 "Increase screen size to display help",
